Dial USSD code in background
Quick one to check the feasibility of dialing USSD services using Telephony Manager.

First, I had to get the list of subscriptions "sims" through subscription manager.
I only care about one currently and that is configured in the Subscription class.

A telephony manager tied to the subscription is created because we want to be able to use just about any subscription later based on subscription name contained in a request.

I was able to successful get this working.