Perform USSD request periodically
A service is defined to schedule recurrent work.

Now, this work is hard-coded.
In the long run, this work will poll a service,
run tasks defined by the service and report results back to the service.

I'ld need to investigate how stable it is to do it this way using the WorkManager as my phone keeps awake the entire time the application is running.
Perhaps using AlarmManager to schedule tasks through the day is more fail safe approach.